feat(button-group): items dropdown component
